% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_depending_cols.R
\name{add_time_depending_cols}
\alias{add_time_depending_cols}
\title{Add time depending columns}
\usage{
add_time_depending_cols(data)
}
\arguments{
\item{data}{tibble returned from load_seasons}
}
\value{
tible with columns HTGS, HTGC, HTGD, HTF, HTLS, HTP, ATGS, ATGC, ATGD, ATF, ATLS, ATP
}
\description{
Add time depending columns
}
\details{
The following time depending columns are added (similar for AT, Away Team):
HTGS: Home Team Goal Scored in the season
HTGC: Home Team Goals Conceded in the season
HTGD: Home Team Goals Difference in the season
HTF: Home Team Form (total points in last 5 games)
HTP: Home Team Points

Note: the columns provide status information prior to the match
}
