% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_numeric.R
\name{all_numeric}
\alias{all_numeric}
\title{All numeric}
\usage{
all_numeric(data)
}
\arguments{
\item{data}{tibble or data.frame}
}
\value{
character vector with all numeric columns
}
\description{
All numeric
}
