% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_predictors.R
\name{all_predictors}
\alias{all_predictors}
\title{Get all predictors in a dataset}
\usage{
all_predictors(data)
}
\arguments{
\item{data}{tibble (see get_modeling_data)}
}
\value{
character vector with possible predictor names
}
\description{
Get all predictors in a dataset
}
