% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternate_vectors.R
\name{alternate_vectors}
\alias{alternate_vectors}
\title{Alternate vectors}
\usage{
alternate_vectors(a, b)
}
\arguments{
\item{a}{vector of any class}

\item{b}{vector of any class}
}
\value{
vector of length 2 * a
}
\description{
Alternate vectors
}
\details{
Vector a and b must be of same length
}
\examples{
alternate_vectors(a = 1:3, b = 11:13)
}
