% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_goals_before_minute_x.R
\name{any_goals_before_minute_x}
\alias{any_goals_before_minute_x}
\title{Any goals before minute x}
\usage{
any_goals_before_minute_x(goals, x)
}
\arguments{
\item{goals}{numeric vector with goalscoring minutes}

\item{x}{integer specifying the minute of interest}
}
\value{
logical, TRUE if any goals where scored before the given minute
}
\description{
Any goals before minute x
}
