% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bet.R
\name{bet_random}
\alias{bet_random}
\title{Bet random}
\usage{
bet_random(data, c = 1)
}
\arguments{
\item{data}{tibble with season data}

\item{c}{integer specifying the seed (default is 1)}
}
\value{
tibble with additional columns
}
\description{
Bet random
}
