% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_earnings.R
\name{calculate_earnings}
\alias{calculate_earnings}
\title{Calculate earnings}
\usage{
calculate_earnings(data)
}
\arguments{
\item{data}{tibble with bets (see bet_model)}
}
\value{
tibble with earning
}
\description{
Calculate earnings
}
