% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_earnings.R
\name{calculate_earnings_by_strategy}
\alias{calculate_earnings_by_strategy}
\title{Calculate earnings by strategy}
\usage{
calculate_earnings_by_strategy(data, f, group_by_date = FALSE)
}
\arguments{
\item{data}{tibble with bets}

\item{f}{list of functions to evaluate (options: bet_home, bet_away, bet_draw and bet_model)}

\item{group_by_date}{logical, indicating if calculations should be made grouped by date (default is FALSE)}
}
\value{
tibble with earnings
}
\description{
Calculate earnings by strategy
}
