% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_avg_market_value.R
\name{convert_million_char_to_numeric}
\alias{convert_million_char_to_numeric}
\title{Convert million char to numeric}
\usage{
convert_million_char_to_numeric(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Convert million char to numeric
}
