% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country}
\alias{country}
\title{Country data}
\format{A data table with 11 rows and 2 variables
\describe{
  \item{id}{Country ID}
  \item{name}{Country name}
}}
\source{
\url{https://www.kaggle.com/hugomathien/soccer}
}
\usage{
country
}
\description{
The 11 countries in the dataset
}
\keyword{datasets}
