% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_health_check.R
\name{data_health_check}
\alias{data_health_check}
\title{Data health check}
\usage{
data_health_check(data, target)
}
\arguments{
\item{data}{tibble, typically modeling data (see get_modeling_data_*)}

\item{target}{optional, character vector with target name}
}
\description{
Data health check
}
