% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_team_model.R
\name{estimate_team_model}
\alias{estimate_team_model}
\title{Estimate team model}
\usage{
estimate_team_model(data)
}
\arguments{
\item{data}{tibble with attack and defence flags (see join_att_and_def_team_flags)}
}
\description{
Estimate team model
}
