% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_performance.R
\name{evaluate_performance}
\alias{evaluate_performance}
\title{Evaluate performance}
\usage{
evaluate_performance(test_data, obj, target)
}
\arguments{
\item{test_data}{test data to evaluate performance based on}

\item{obj}{glm object}

\item{target}{character string with response name}
}
\description{
Evaluate performance
}
