% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_games_grid.R
\name{expand_games_grid}
\alias{expand_games_grid}
\title{Expand games grid}
\usage{
expand_games_grid(team_names, league, season)
}
\arguments{
\item{team_names}{charater vector with team names (see `get_team_names_from_world_football`)}

\item{league}{character string with league name (for instance "premier-league")}

\item{season}{character string with season(for istance "2021-2022")}
}
\value{
tibble with grid consisting of all games possible from `team_names`
}
\description{
Expand games grid
}
