% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{football}
\alias{football}
\alias{football-package}
\title{European professional football data from Kaggle}
\description{
European professional football data from Kaggle for the
  R programming course. All data taken from
  https://www.kaggle.com/hugomathien/soccer
}
\section{Datasets}{

\describe{
  \item{country}{}
  \item{football_match}{}
  \item{league}{}
  \item{player}{}
  \item{player_attributes}{}
  \item{team}{}
  \item{team_attributes}{}
}
}

