% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{football.teams}
\alias{football.teams}
\title{Merged Data}
\format{A data frame with 17 rows and 12 variables.}
\source{
1.\url{https://m.qtx.com/dbsoccer/stading_92_2018.html}
2.\url{https://api.football-data.org/v2/teams/86/matches?status=SCHEDULED.}
}
\usage{
football.teams
}
\description{
This is the merged data set from web scraping and the collection data from API.
}
\keyword{datasets}
