% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{football_match}
\alias{football_match}
\title{Football Match}
\format{A data table with 25979 rows and 115 variables
\describe{
\item{id}{}
\item{country_id}{}
\item{league_id}{}
\item{season}{}
\item{stage}{}
\item{date}{}
\item{match_api_id}{}
\item{home_team_api_id}{}
\item{away_team_api_id}{}
\item{home_team_goal}{}
\item{away_team_goal}{}
\item{home_player_X1}{}
\item{home_player_X2}{}
\item{home_player_X3}{}
\item{home_player_X4}{}
\item{home_player_X5}{}
\item{home_player_X6}{}
\item{home_player_X7}{}
\item{home_player_X8}{}
\item{home_player_X9}{}
\item{home_player_X10}{}
\item{home_player_X11}{}
\item{away_player_X1}{}
\item{away_player_X2}{}
\item{away_player_X3}{}
\item{away_player_X4}{}
\item{away_player_X5}{}
\item{away_player_X6}{}
\item{away_player_X7}{}
\item{away_player_X8}{}
\item{away_player_X9}{}
\item{away_player_X10}{}
\item{away_player_X11}{}
\item{home_player_Y1}{}
\item{home_player_Y2}{}
\item{home_player_Y3}{}
\item{home_player_Y4}{}
\item{home_player_Y5}{}
\item{home_player_Y6}{}
\item{home_player_Y7}{}
\item{home_player_Y8}{}
\item{home_player_Y9}{}
\item{home_player_Y10}{}
\item{home_player_Y11}{}
\item{away_player_Y1}{}
\item{away_player_Y2}{}
\item{away_player_Y3}{}
\item{away_player_Y4}{}
\item{away_player_Y5}{}
\item{away_player_Y6}{}
\item{away_player_Y7}{}
\item{away_player_Y8}{}
\item{away_player_Y9}{}
\item{away_player_Y10}{}
\item{away_player_Y11}{}
\item{home_player_1}{}
\item{home_player_2}{}
\item{home_player_3}{}
\item{home_player_4}{}
\item{home_player_5}{}
\item{home_player_6}{}
\item{home_player_7}{}
\item{home_player_8}{}
\item{home_player_9}{}
\item{home_player_10}{}
\item{home_player_11}{}
\item{away_player_1}{}
\item{away_player_2}{}
\item{away_player_3}{}
\item{away_player_4}{}
\item{away_player_5}{}
\item{away_player_6}{}
\item{away_player_7}{}
\item{away_player_8}{}
\item{away_player_9}{}
\item{away_player_10}{}
\item{away_player_11}{}
\item{goal}{}
\item{shoton}{}
\item{shotoff}{}
\item{foulcommit}{}
\item{card}{}
\item{cross}{}
\item{corner}{}
\item{possession}{}
\item{B365H}{}
\item{B365D}{}
\item{B365A}{}
\item{BWH}{}
\item{BWD}{}
\item{BWA}{}
\item{IWH}{}
\item{IWD}{}
\item{IWA}{}
\item{LBH}{}
\item{LBD}{}
\item{LBA}{}
\item{PSH}{}
\item{PSD}{}
\item{PSA}{}
\item{WHH}{}
\item{WHD}{}
\item{WHA}{}
\item{SJH}{}
\item{SJD}{}
\item{SJA}{}
\item{VCH}{}
\item{VCD}{}
\item{VCA}{}
\item{GBH}{}
\item{GBD}{}
\item{GBA}{}
\item{BSH}{}
\item{BSD}{}
\item{BSA}{}
}}
\source{
\url{https://www.kaggle.com/hugomathien/soccer}
}
\usage{
football_match
}
\description{
Football Match
}
\keyword{datasets}
