% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abbrev_team_names.R
\name{get_abbrev_team_names}
\alias{get_abbrev_team_names}
\title{Get abbreviated team names}
\usage{
get_abbrev_team_names(x)
}
\arguments{
\item{x}{character vector with team names}
}
\description{
Get abbreviated team names
}
