% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assist.R
\name{get_assists}
\alias{get_assists}
\title{Get assists from match report}
\usage{
get_assists(df)
}
\arguments{
\item{df}{data.frame with goals scored (df_goals obtained from `get_match_report`)}
}
\value{
character vector with assist names
}
\description{
Get assists from match report
}
