% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_avg_market_value.R
\name{get_avg_market_value}
\alias{get_avg_market_value}
\title{Get average market value by club}
\usage{
get_avg_market_value(season)
}
\arguments{
\item{season}{season as integer (for instance 2020. 2020 corresponds to season 2021)}
}
\description{
Get average market value by club
}
