% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champions_league.R
\name{get_champions_league_goal_scoring_minute}
\alias{get_champions_league_goal_scoring_minute}
\title{Get champions league goal socring minute}
\usage{
get_champions_league_goal_scoring_minute(df_games)
}
\arguments{
\item{df_games}{tibble with HomeTeam, AwayTeam, Season and Stage}
}
\description{
Get champions league goal socring minute
}
