% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_champions_league_results.R
\name{get_champions_league_results}
\alias{get_champions_league_results}
\title{Get champions league results}
\usage{
get_champions_league_results(season)
}
\arguments{
\item{season}{character string with season (for instance "2021-2022")}
}
\value{
tibble with results
}
\description{
Get champions league results
}
