% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country_from_league.R
\name{get_country_from_league}
\alias{get_country_from_league}
\title{Get country from league}
\usage{
get_country_from_league(league)
}
\arguments{
\item{league}{league name (character string, for instance "premier-league" or "primera-division")}
}
\description{
Get country from league
}
