% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_futbin_url.R
\name{get_futbin_url}
\alias{get_futbin_url}
\title{Get futbin url}
\usage{
get_futbin_url(year, league_name, club_name, page_number)
}
\arguments{
\item{year}{year (integer), for instance 20}

\item{page_number}{page number (integer)}
}
\value{
character string
}
\description{
Get futbin url
}
