% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_goalscoring_type.R
\name{get_goalscoring_type}
\alias{get_goalscoring_type}
\title{Get goalscoring type}
\usage{
get_goalscoring_type(df)
}
\arguments{
\item{df}{tibble with goal scored (df_goals obtained from get_match_report)}
}
\description{
Get goalscoring type
}
