% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{get_lambda}
\alias{get_lambda}
\title{Get lambda}
\usage{
get_lambda(data, obj)
}
\arguments{
\item{data}{tibble with season data (see load_seasons)}

\item{obj}{glm object}
}
\value{
input tibble is returned with additional two columns: lambda_home_team and lambda_away_team
}
\description{
Get lambda
}
