% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_league_table.R
\name{get_league_table}
\alias{get_league_table}
\title{Get league table}
\usage{
get_league_table(league, season, round = 1)
}
\arguments{
\item{league}{league name (for instance "premier-league")}

\item{season}{season (for insance "2021-2022")}

\item{round}{round (integer between 1 and 38) indicating game round (default is 1)}
}
\description{
Get league table
}
