% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_minutes_goals_scored.R
\name{get_minutes_goals_scored}
\alias{get_minutes_goals_scored}
\title{Get minutes goals scored}
\usage{
get_minutes_goals_scored(df)
}
\arguments{
\item{df}{data.frame with columns X1 (= the score) and X2 (= string with goal scorer, minute, how the goal was scored and assist)}
}
\value{
integer vector with minutes
}
\description{
Get minutes goals scored
}
