% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{get_model_bets}
\alias{get_model_bets}
\title{Get model bets}
\usage{
get_model_bets(data, obj, verbose = FALSE)
}
\arguments{
\item{data}{tibble with season data}

\item{obj}{glm object}

\item{verbose}{logical indicating if loop counter should be on (TRUE) or off (FALSE)}
}
\value{
tibble with season data and additional column related to probability, odds, ratio and bet
}
\description{
Get model bets
}
