% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modeling_data.R
\name{get_modeling_data}
\alias{get_modeling_data}
\title{Get modeling data}
\usage{
get_modeling_data(data)
}
\arguments{
\item{data}{tibble obtained from `join_fifa_ratings` using data from `load_seasons`}
}
\description{
Get modeling data
}
