% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modeling_data.R
\name{get_modeling_data_simple}
\alias{get_modeling_data_simple}
\title{Get modeling data}
\usage{
get_modeling_data_simple(data)
}
\arguments{
\item{data}{tibble obtained from `load_seasons`}
}
\description{
Get modeling data
}
