% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{get_prob}
\alias{get_prob}
\title{Get probability for home, draw and away win}
\usage{
get_prob(lambda_home_team, lambda_away_team)
}
\arguments{
\item{lambda_home_team}{numeric value specifying lambda (= expected goals pr. game) for the home team}

\item{lambda_away_team}{numeric value specifying lambda (= expected goals pr. game) for the away team}

\item{obj}{glm object (use glm with join_att_and_def_team_flags)}
}
\description{
Get probability for home, draw and away win
}
