% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_formula.R
\name{get_team_formula}
\alias{get_team_formula}
\title{Get team formula}
\usage{
get_team_formula(data, intercept = FALSE, home = FALSE)
}
\arguments{
\item{data}{tibble with modeling data and attack and defence flags (see join_att_and_def_team_flags)}

\item{home}{logical indicating if Home variable should be included in the formula}
}
\value{
formula with all attack and defence terms for a design matrix with full rang
}
\description{
Get team formula
}
