% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_names.R
\name{get_team_names}
\alias{get_team_names}
\title{Get team names}
\usage{
get_team_names(data)
}
\arguments{
\item{data}{tibble obtained from `load_seasons`}
}
\value{
character vector with team names in data
}
\description{
Get team names
}
