% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_names_from_world_football.R
\name{get_team_names_from_world_football}
\alias{get_team_names_from_world_football}
\title{Get team names from worldfootball}
\usage{
get_team_names_from_world_football(league, season)
}
\arguments{
\item{league}{league name (for instance "premier-league")}

\item{season}{season (for insance "2021-2022")}
}
\description{
Get team names from worldfootball
}
