% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seasons.R
\name{get_url}
\alias{get_url}
\title{Load one season
Get URL for loading data}
\usage{
get_url(league, season)
}
\arguments{
\item{league}{character string with league name (Premier League or La Liga)}

\item{season}{character string indicating season, for instance "2122"}
}
\description{
Load one season
Get URL for loading data
}
