% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_for_match_report.R
\name{get_url_for_match_report}
\alias{get_url_for_match_report}
\title{Get url for match report}
\usage{
get_url_for_match_report(league, season, home_team, away_team)
}
\arguments{
\item{league}{league name (for instance "premier-league")}

\item{season}{season (for insance "2021-2022")}

\item{home_team}{name of home team (for instance "arsenal-fc")}

\item{away_team}{name of away team (for instance "manchester-united")}
}
\description{
Get url for match report
}
