% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_odds.R
\name{implied_odds}
\alias{implied_odds}
\title{Implied odds}
\usage{
implied_odds(p)
}
\arguments{
\item{p}{probabilities (numeric values)}
}
\value{
numeric vector of same dimension as p
}
\description{
Implied odds
}
