% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_odds.R
\name{implied_probability}
\alias{implied_probability}
\title{Implied probablity}
\usage{
implied_probability(o, ignore_margen = FALSE, margen = 1.05, n_outcomes = 2)
}
\arguments{
\item{o}{odds (numeric value)}
}
\description{
Implied probablity
}
