% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_att_and_def_flags.R
\name{join_att_and_def_flags}
\alias{join_att_and_def_flags}
\title{Join attack and defence flags}
\usage{
join_att_and_def_flags(data, verbose = FALSE)
}
\arguments{
\item{data}{tibble with modeling data (see get_modeling_data_*)}

\item{verbose}{logical indicating if loop counter should be on (default is FALSE)}
}
\description{
Join attack and defence flags
}
