% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_fifa_ratings.R
\name{join_fifa_ratings}
\alias{join_fifa_ratings}
\title{Join fifa ratings}
\usage{
join_fifa_ratings(data)
}
\arguments{
\item{data}{tibble obtained from `load_seasons`}
}
\description{
Join fifa ratings
}
