% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_transfermarkt_data.R
\name{join_transfermarkt_data}
\alias{join_transfermarkt_data}
\title{Join transfermarkt data}
\usage{
join_transfermarkt_data(data, season)
}
\arguments{
\item{data}{tibble with season data (see get_modeling data)}

\item{season}{integer (for instance 2020 which covers 2021 season)}
}
\description{
Join transfermarkt data
}
