% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{league}
\alias{league}
\title{League}
\format{A data table with 11 rows and 3 variables
\describe{
\item{id}{Same as the \code{id} column from \code{country}}
\item{country_id}{Matches the \code{id} column from \code{country}}
\item{name}{Full name of the national league}
}}
\source{
\url{https://www.kaggle.com/hugomathien/soccer}
}
\usage{
league
}
\description{
League
}
\keyword{datasets}
