% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_estimate_and_place_bets.R
\name{load_estimate_and_place_bets}
\alias{load_estimate_and_place_bets}
\title{Load estimate and place bets}
\usage{
load_estimate_and_place_bets(league = "Premier League", from = from, to = to)
}
\arguments{
\item{from}{integer with from year (2 digits, 18 meaning 2018)}

\item{to}{integer with to year (2 digits, 19 meaning 2019)}

\item{leauge}{character string with league name (default is "Premier League")}
}
\value{

}
\description{
Load estimate and place bets
}
