% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fifa_ratings.R
\name{load_fifa_ratings}
\alias{load_fifa_ratings}
\title{Load fifa ratings}
\usage{
load_fifa_ratings(
  year = 20,
  league_name = "Premier League",
  club_name = "Liverpool",
  max_pages = 10
)
}
\arguments{
\item{year}{year as integer (for instance 20)}

\item{league_name}{league name (for instance "Premier League")}

\item{club_name}{club name (for instance "Liverpool". See `get_team_names`}

\item{nax_pages}{maximal number of pages to load}
}
\description{
Load fifa ratings
}
