% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seasons.R
\name{load_one_season}
\alias{load_one_season}
\title{Load one season}
\usage{
load_one_season(season)
}
\arguments{
\item{season}{character string indicating season, for instance "2122"}
}
\description{
Load one season
}
