% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seasons.R
\name{load_seasons}
\alias{load_seasons}
\title{Load seasons}
\usage{
load_seasons(from, to, verbose = TRUE)
}
\arguments{
\item{from}{from year (integer), for instance 18}

\item{to}{to year (integer), for instance 22}

\item{verbose}{logical indicating if looper counter should be on (TRUE) or off (FALSE)}
}
\value{
tibble with football results
}
\description{
Load seasons
}
