% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_squared_using_mae.R
\name{mae}
\alias{mae}
\title{Mean absolute error}
\usage{
mae(x, y)
}
\arguments{
\item{x}{numeric values, typically the predictions}

\item{y}{numeric values, typiicaly the response variable}
}
\description{
Mean absolute error
}
