% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_club_name_to_number.R
\name{map_club_name_to_number}
\alias{map_club_name_to_number}
\title{Map club name to number}
\usage{
map_club_name_to_number(data, club_name)
}
\arguments{
\item{data}{tibble with historical football results. See `load_seasons`}

\item{club_name}{character string with club name (for instance "Liverpool")}
}
\description{
Map club name to number
}
\details{
The number is found by sorting all club names and picking the index corresponding to `club_name`
For this reason `data` argument is needed
}
