% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fifa_team_name_to_uk_data.R
\name{map_fifa_team_name_to_uk_data}
\alias{map_fifa_team_name_to_uk_data}
\title{Map fifa team name to uk data}
\usage{
map_fifa_team_name_to_uk_data(x)
}
\description{
Map fifa team name to uk data
}
