% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_league_name_to_number.R
\name{map_league_name_to_number}
\alias{map_league_name_to_number}
\title{Map league to number}
\usage{
map_league_name_to_number(league_name)
}
\arguments{
\item{league_name}{character string with leage name (for instance "Premier League")}
}
\value{
integer
}
\description{
Map league to number
}
