% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_number_to_club_name.R
\name{map_number_to_club_name}
\alias{map_number_to_club_name}
\title{Map number to club name}
\usage{
map_number_to_club_name(year, league_name, start, finish)
}
\arguments{
\item{year}{year as integer (20)}

\item{league_name}{league name ("Premier League")}

\item{start}{integer indicating first page to load}

\item{finish}{integer indicating last page to load}
}
\description{
Map number to club name
}
