% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_transfermarkt_to_season_names.R
\name{map_transfermarkt_to_season_names}
\alias{map_transfermarkt_to_season_names}
\title{Map transfermarkt names to season names}
\usage{
map_transfermarkt_to_season_names(names_transfermarkt, names_season)
}
\description{
Map transfermarkt names to season names
}
