% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_year_to_season.R
\name{map_year_to_season}
\alias{map_year_to_season}
\title{Map year to season}
\usage{
map_year_to_season(year)
}
\arguments{
\item{year}{year as integer (for instance 20)}
}
\description{
Map year to season
}
