% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{matchday_team_attributes}
\alias{matchday_team_attributes}
\title{A view to see matchday team attributes}
\usage{
matchday_team_attributes()
}
\value{
A data.table with all the home and away team attributes.
}
\description{
A view to see matchday team attributes
}
\examples{
matchday_team_attributes()
}
