% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champions_league.R
\name{modify_team_name_for_worldfootball}
\alias{modify_team_name_for_worldfootball}
\title{Modify team name for worldfootball}
\usage{
modify_team_name_for_worldfootball(x)
}
\arguments{
\item{x}{character string (typically this function is used with team names)}
}
\description{
Modify team name for worldfootball
}
