% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_champions_league_results.R
\name{mutate_and_select_champions_league_cols}
\alias{mutate_and_select_champions_league_cols}
\title{Mutate and select champions league cols}
\usage{
mutate_and_select_champions_league_cols(data, stage)
}
\arguments{
\item{data}{tibble to mutate and select desired columns based on the given stage}

\item{stage}{character string with stage name}
}
\description{
Mutate and select champions league cols
}
