% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{player_to_match}
\alias{player_to_match}
\title{Return the match_api_id for all matches starting a particular player}
\usage{
player_to_match(player_api_id)
}
\arguments{
\item{player_api_id}{The \code{player_api_id} from the \code{player} dataset.}
}
\value{
A data.table of all the matches where the player started.
Columns are:
\itemize{
\item{match_api_id}
\item{home_team_api_id}
\item{away_team_api_id}
\item{date: Match date}
\item{at_home: (Boolean) Did the player start for the home side?}
}
}
\description{
Return the match_api_id for all matches starting a particular player
}
\examples{
# All Lionel Messi's matches
player_to_match(30981)
}
