% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_earnings_by_day.R
\name{plot_earnings_by_day}
\alias{plot_earnings_by_day}
\title{Plot earnings by day}
\usage{
plot_earnings_by_day(data)
}
\arguments{
\item{data}{tibble with bets (see bet_model)}
}
\value{
plotly object
}
\description{
Plot earnings by day
}
