% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_earnings_by_day_and_model.R
\name{plot_earnings_by_day_and_model}
\alias{plot_earnings_by_day_and_model}
\title{Plot earnings by day and model}
\usage{
plot_earnings_by_day_and_model(data, title = "")
}
\arguments{
\item{data}{tibble with bets}

\item{title}{character string with plot title}
}
\value{
plotly object
}
\description{
Plot earnings by day and model
}
