% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_performance.R
\name{plot_performance}
\alias{plot_performance}
\title{Graph the Team Performance of Different Teams.}
\usage{
plot_performance(data, perfrm_vars, teams, all = FALSE)
}
\arguments{
\item{data}{the data set of merged data}

\item{perfrm_vars}{the performance variables that want to compare}

\item{teams}{the team names of given teams}

\item{all}{If all is true, it will give the performance comparison of all teams.
The default value of all is FALSE.}
}
\description{
This function plots the team performance of different teams from England.
}
\examples{
# Example: Graph the Performance of Different Teams

teams <- c("Liverpool", "Chelsea", "Arsenal", "Chelsea", "everton", "fulham")
plot_performance(football.teams, teams, perfrm_vars = c("points", "won", "loss"))

}
