% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_squared_using_mae.R
\name{r_squared_using_mae}
\alias{r_squared_using_mae}
\title{R squared calculated using mean absolute error (MAE) rather than likelihood function}
\usage{
r_squared_using_mae(obj, data, target = "Goals")
}
\arguments{
\item{obj}{glm object}

\item{data}{optional, data to evaluate r squared on (default is training data taken from the glm object)}

\item{target}{character string specifying the target/response variable}
}
\description{
R squared calculated using mean absolute error (MAE) rather than likelihood function
}
