% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_html_team.R
\name{read_html_team}
\alias{read_html_team}
\title{Get the Basic Information of Football Teams from England}
\usage{
read_html_team(teams, all = FALSE)
}
\arguments{
\item{teams}{the team names of given teams}

\item{all}{If all is true, it will return the basic information of all.The default value
of all is FALSE.}
}
\value{
the basic information of football teams,
including country, home, city and fundation time.
}
\description{
The function gets the basic information of a football team from England.
The data is got from https://m.qtx.com/dbsoccer/stading_92_2018.html using web sraping.
The baisc information includes country, home, city and fundation time.
}
\examples{
# Example 1: Get the basic information of Liverpool, Chelsea
info_sample <- read_html_team(teams = c("Liverpool", "Chelsea"))
print(info_sample)
# Example 2: Get the basic information of all football teams from England
info_all <- read_html_team(all = TRUE)
print(info_all)

}
