% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champions_league.R
\name{remove_acutes}
\alias{remove_acutes}
\title{Remove acutes}
\usage{
remove_acutes(x)
}
\arguments{
\item{x}{character string (typically this function is used with team names)}
}
\description{
Remove acutes
}
