% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_distr_cols.R
\name{rename_distr_cols}
\alias{rename_distr_cols}
\title{Rename distribution columns}
\usage{
rename_distr_cols(data)
}
\arguments{
\item{data}{tibble (typically modeling data, see get_modeling_data_*)}
}
\description{
Rename distribution columns
}
