% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seasons.R
\name{season_seq}
\alias{season_seq}
\title{Season sequence}
\usage{
season_seq(from, to)
}
\arguments{
\item{from}{from year (integer), for instance 18}

\item{to}{to year (integer), for instance 22}
}
\description{
Season sequence
}
