% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_estimation.R
\name{stepwise_estimation}
\alias{stepwise_estimation}
\title{Stepwise esimation}
\usage{
stepwise_estimation(data, alpha = 0.05, max_iteration = 40, intercept = TRUE)
}
\arguments{
\item{data}{tibble with attack and defence flags (see join_att_and_def_team_flags)}

\item{alpha}{signifiance level (default is 0.05)}

\item{max_iteration}{maximum number of steps}

\item{intercept}{intercept, logical indicating an intercept will be estimated}
}
\description{
Stepwise esimation
}
