% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_football.R
\name{summary_football}
\alias{summary_football}
\title{The Basic Statistics of Football Team's Performance}
\usage{
summary_football(data, perfrm_vars, teams, all = FALSE)
}
\arguments{
\item{data}{the data set of merged data}

\item{perfrm_vars}{the performance variables}

\item{teams}{the team names of given teams}

\item{all}{If all is true, it will give the summary statistics of all teams.
The default value of all is FALSE.}
}
\value{
the basic statistics of football team Performance
}
\description{
This function returns the basic statistics of football team's performance variables
}
\examples{
# Example get the summary statistics of points, won, and loss of all teams
summary_football(data = football.teams, perfrm_vars = c("points", "won", "loss"), all = TRUE)
}
