% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{team}
\alias{team}
\title{Team data}
\format{A data table with 299 rows and 5 variables
\describe{
  \item{id}{}
  \item{team_api_id}{}
  \item{team_fifa_api_id}{}
  \item{team_long_name}{}
  \item{team_short_name}{}
}}
\source{
\url{https://www.kaggle.com/hugomathien/soccer}
}
\usage{
team
}
\description{
Team data
}
\keyword{datasets}
