% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{two_way_interaction_data}
\alias{two_way_interaction_data}
\title{Two-way interactions data}
\usage{
two_way_interaction_data(training_data, num_factor_length = 5L)
}
\arguments{
\item{training_data}{A data.table with the features of interest}

\item{num_factor_length}{The number of cuts to make to numeric cols}
}
\value{
A list: data.table and indexes
}
\description{
Two-way interactions data
}
