% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_season.R
\name{validate_season}
\alias{validate_season}
\title{Validate season}
\usage{
validate_season(season)
}
\arguments{
\item{season}{character string with season (for instance "2021-2022")}
}
\description{
Validate season
}
