% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorized_minutes_goals_scored.R
\name{vectorized_minutes_goals_scored}
\alias{vectorized_minutes_goals_scored}
\title{Vectorized minutes goals scored}
\usage{
vectorized_minutes_goals_scored(df)
}
\arguments{
\item{df}{tibble with HomeTeam and AwayTeam (see `expand_games_grid`)}
}
\value{
list with minutes goals scored for each game
}
\description{
Vectorized minutes goals scored
}
