% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{vectorized_prob}
\alias{vectorized_prob}
\title{Vectorized probabilities}
\usage{
vectorized_prob(data, obj, verbose)
}
\arguments{
\item{data}{tibble with season data (see load_seasons)}

\item{obj}{glm obj}

\item{verbose}{logical indicating if loop counter should be on (TRUE) or off (FALSE)}
}
\value{
tibble with input data and 6 additional columns (probabilities and ratios for home, draw and away win)
}
\description{
Vectorized probabilities
}
