% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_factor_analysis.R
\name{single_factor_analysis}
\alias{single_factor_analysis}
\title{Single factor analysis}
\usage{
single_factor_analysis(data, predictors, verbose = TRUE)
}
\arguments{
\item{data}{tibble with modeling data (see get_modeling_data_*)}

\item{predictors}{character vector with predictor names}

\item{verbose}{logical indicating if loop counter should be on (default is TRUE)}
}
\description{
Single factor analysis
}
